requirements.txt
Django==1.11.5
django-bootstrap3==9.0.0
django-safedelete==0.4.2
psycopg2==2.7.3.1
pytz==2017.2
